Withdrawal of unnecessary movings
