+++ pystiche_papers/johnson_alahi_li_2016/_data.py
@@ -1,8 +1,11 @@
-from typing import Optional, Sized
+from typing import Optional, Sized, Tuple
 from urllib.parse import urljoin
 
+import torch
 from torch.utils.data import DataLoader, Dataset, Sampler
 
+import pystiche.image.transforms.functional as F
+from pystiche import image
 from pystiche.data import (
     DownloadableImage,
     DownloadableImageCollection,
@@ -11,7 +14,7 @@ from pystiche.data import (
 from pystiche.image import transforms
 
 from ..data.utils import FiniteCycleBatchSampler
-from ..utils.transforms import OptionalGrayscaleToFakegrayscale, TopLeftCropToMultiple
+from ..utils.transforms import OptionalGrayscaleToFakegrayscale
 
 __all__ = [
     "content_transform",
@@ -26,6 +29,20 @@ __all__ = [
 def content_transform(
     edge_size: int = 256, multiple: int = 16, impl_params: bool = True,
 ) -> transforms.ComposedTransform:
+    class TopLeftCropToMultiple(transforms.Transform):
+        def __init__(self, multiple: int):
+            super().__init__()
+            self.multiple = multiple
+
+        def calculate_size(self, input_image: torch.Tensor) -> Tuple[int, int]:
+            old_height, old_width = image.extract_image_size(input_image)
+            new_height = old_height - old_height % self.multiple
+            new_width = old_width - old_width % self.multiple
+            return new_height, new_width
+
+        def forward(self, input_image: torch.Tensor) -> torch.Tensor:
+            size = self.calculate_size(input_image)
+            return F.top_left_crop(input_image, size)
 
     transforms_ = [
         TopLeftCropToMultiple(multiple),
