+++ pystiche_papers/utils/modules.py
@@ -1,19 +1,12 @@
-from collections import OrderedDict
 from typing import Any, Dict, Optional, Union, cast
 
 import torch
 from torch import nn
 
-from pystiche import meta as meta_
-
-from ..utils import join_channelwise
-
 __all__ = [
     "Identity",
     "ResidualBlock",
     "SequentialWithOutChannels",
-    "AddNoiseChannels",
-    "HourGlassBlock",
 ]
 
 
@@ -48,36 +41,3 @@ class SequentialWithOutChannels(nn.Sequential):
         self.out_channels = cast(Dict[str, nn.Module], self._modules)[
             out_channel_name
         ].out_channels
-
-
-class AddNoiseChannels(nn.Module):
-    def __init__(
-        self, in_channels: int, num_noise_channels: int = 3,
-    ):
-        super().__init__()
-        self.num_noise_channels = num_noise_channels
-        self.in_channels = in_channels
-        self.out_channels = in_channels + num_noise_channels
-
-    def forward(self, input: torch.Tensor) -> torch.Tensor:
-        size = self._extract_size(input)
-        meta = meta_.tensor_meta(input)
-        noise = torch.rand(size, **meta)
-        return join_channelwise(input, noise)
-
-    def _extract_size(self, input: torch.Tensor) -> torch.Size:
-        size = list(input.size())
-        size[1] = self.num_noise_channels
-        return torch.Size(size)
-
-
-class HourGlassBlock(SequentialWithOutChannels):
-    def __init__(
-        self, downsample: nn.Module, intermediate: nn.Module, upsample: nn.Module
-    ):
-        modules = (
-            ("down", downsample),
-            ("intermediate", intermediate),
-            ("up", upsample),
-        )
-        super().__init__(OrderedDict(modules), out_channel_name="intermediate")
